Main

Description

Módulo para a realização da Tarefa 5 de LI1 em 2017/18.

Documentation

game :: Jogo

game é um Jogo definido por: Jogo{mapa = (Mapa ((2,1),Este) tab),pista = prop,carros = [carro1,carro2,carro3,carro4],nitros = [0,0,0,0],historico = [[],[],[],[]]}

prop :: Propriedades

prop são Propriedades definidas por: Propriedades {k_atrito = 0.0,k_pneus = 0.0,k_acel = 0.0,k_peso = 0.0,k_nitro = 0.0,k_roda = 0.0}

tab :: Tabuleiro

tab é um Tabuleiro definido por: [[Peca Lava 0,Peca Lava 0,Peca Lava 0,Peca Lava 0,Peca Lava 0,Peca Lava 0,Peca Lava 0],[Peca Lava 0,Peca (Curva Norte) 0,Peca Recta 0,Peca Recta 0,Peca Recta 0,Peca (Curva Este) 0,Peca Lava 0],[Peca Lava 0,Peca Recta 1,Peca Lava 0,Peca Lava 0,Peca Lava 0,Peca Recta 0,Peca Lava 0],[Peca Lava 0,Peca Recta 1,Peca Lava 0,Peca (Curva Norte) 2,Peca Recta 1,Peca (Curva Sul) 1,Peca Lava 0],[Peca Lava 0,Peca Recta 2,Peca Lava 0,Peca Recta 2,Peca Lava 0,Peca Lava 0,Peca Lava 0],[Peca Lava 0,Peca (Curva Oeste) 2,Peca Recta 2,Peca (Curva Sul) 2,Peca Lava 0,Peca Lava 0,Peca Lava 0],[Peca Lava 0,Peca Lava 0,Peca Lava 0,Peca Lava 0,Peca Lava 0,Peca Lava 0,Peca Lava 0]]

carro1 :: Carro

carro1 é um Carro definido por: Carro {posicao = (2.5,1.1) ,direcao = 0,velocidade = (1,0)}

carro2 :: Carro

carro2 é um Carro definido por: Carro {posicao = (2.5,1.35) ,direcao = 0,velocidade = (1,0)}

carro3 :: Carro

carro3 é um Carro definido por: Carro {posicao = (2.5,1.6) ,direcao = 0,velocidade = (1,0)}

carro4 :: Carro

carro4 é um Carro definido por: Carro {posicao = (2.5,1.85) ,direcao = 0,velocidade = (1,0)}

type Estado = (Jogo, Tabuleiro, Picture, Picture, Picture, Picture, Picture, Picture, Picture, Picture, Picture, Picture, Picture, Picture, Picture, Picture, Float)

Estado é definido por: (Jogo, Tabuleiro, Picture, Picture, Picture, Picture, Picture, Picture,Picture, Picture, Picture, Picture, Picture, Picture, Picture, Picture, Float)

type Estado2 = ([Peca], Picture, Picture, Picture, Picture, Picture, Picture, Picture, Picture, Picture, Picture)

Estado2 é definido por: ([Peca], Picture, Picture, Picture, Picture, Picture,Picture, Picture, Picture, Picture, Picture)

getsize :: Estado -> Float

A função getsize recebe um Estado e devolve um Float com o seu número de colunas.

estadoInicial :: IO Estado

A função estadoInicial gera o estado inicial do Estado.

desenhaEstado :: Estado -> Picture

A função desenhaEstado recebe um Estado e transforma-o numa Picture, ou seja, desenha o estado.

time :: Estado -> Float

A função time recebe um Estado e devolve o tempo remanescente de jogo.

desenhaCar :: Carro -> Int -> Float -> Picture -> Picture

A função desenhaCar desenha o Carro que recebe na respectiva posição do tabuleiro.

desenhaCarro :: Estado -> [Picture]

A função desenhaCarro desenha os Carros existentes no Estado através do uso da função desenhaCar.

auxdesenhaEstado :: Estado -> [Picture]

A função auxdesenhaEstado devolve uma lista de Picture com as Pecas do Tabuleiro.

desenhaMapa :: [Picture] -> Float -> Float -> Float -> Float -> [Picture]

A função desenhaMapa recebe a lista gerada pela auxdesenhaEstado e coloca-as nas posições pretendidas.

reageEvento :: Event -> Estado -> Estado

A função reageEvento recebe um Event e um Estado, retornando o novo Estado, consoante a alteração vinda do evento.

reageTempo :: Float -> Estado -> Estado

A função reageTempo retira o Float que recebe ao Tempo existente no Estado.

fr :: Int

fr são os frames por segundo.

dm :: Display

dm é a dimensão da janela do jogo.

main :: IO ()

Função principal usada para animar um jogo completo. Compilar com o GHC.

a1 :: [Acao]

a1 é uma '[Acao]' definido por: [acao4,acao2,acao2,acao2].

a2 :: [Acao]

a2 é uma '[Acao]' definido por: [acao5,acao2,acao2,acao2].

a3 :: [Acao]

a3 é uma '[Acao]' definido por: [acao6,acao2,acao2,acao2].

a4 :: [Acao]

a4 é uma '[Acao]' definido por: [acao7,acao2,acao2,acao2].

acao5 :: Acao

acao5 é uma Acao definido por: Acao {acelerar = False, travar = True, esquerda= False, direita= False, nitro = Nothing}

acao6 :: Acao

acao6 é uma Acao definido por: Acao {acelerar = False, travar = False, esquerda= True, direita= False, nitro = Nothing}

acao7 :: Acao

acao7 é uma Acao definido por: Acao {acelerar = False, travar = False, esquerda= False, direita= True, nitro = Nothing}

atualizaMovimenta :: Tempo -> Jogo -> [Acao] -> Jogo

A função atualizaMovimenta atualiza o estado do jogo, através do uso das funções principais das Tarefas 3 e 4.

atualizar :: Tempo -> Jogo -> [Acao] -> Jogo

A função atualizar aplica a função atualiza da Tarefa 4 a todos os jogadores.

movimentar :: Tempo -> Jogo -> Jogo

A função movimentar aplica a função movimenta da Tarefa 3 a todos os jogadores.

fromJust2 :: Maybe Carro -> Carro

A função fromJust2 recebe um Maybe Carro e devolve um Carro, quando o input é diferente de Nothing.

revive :: Carro -> Posicao -> Carro

A função revive recebe um Carro e uma Posicao, colocando o Carro na mesma, com angulo 0 e velocidade (0,0).

posit :: Jogo -> Int -> Posicao

A função posit tira a última Posicao de um certo Carro.

tabuleiro :: Jogo -> Tabuleiro

A função tabuleiro devolve o Tabuleiro a partir de um certo Jogo.