Tarefa2_2017li1g86

Description

Módulo que implementa as funções da Tarefa 2 de Laboratórios de Informática 1.

Documentation

testesT2 :: [Tabuleiro]

testesT2 é um lista de Tabuleiro definido por: [(tab m), (tab m2)]

m :: Mapa

m é um Mapa definido por: Mapa ((2,1),Este) [[Peca Lava 0,Peca Lava 0,Peca Lava 0,Peca Lava 0,Peca Lava 0],[Peca Lava 0,Peca (Curva Norte) 0,Peca Recta 0,Peca (Curva Este) 0,Peca Lava 0],[Peca Lava 0,Peca (Rampa Sul) 0,Peca Lava 0,Peca (Rampa Sul) 0,Peca Lava 0],[Peca Lava 0,Peca (Curva Oeste) 1,Peca Recta 1,Peca (Curva Sul) 1,Peca Lava 0],[Peca Lava 0,Peca Lava 0,Peca Lava 0,Peca Lava 0,Peca Lava 0]]

m2 :: Mapa

m2 é um Mapa definido por: Mapa ((2,1),Este) [[Peca Lava 0,Peca Lava 0,Peca Lava 0,Peca Lava 0,Peca Lava 0],[Peca Lava 0,Peca (Curva Norte) 0,Peca Lava 1,Peca (Curva Este) 0,Peca Lava 0],[Peca Lava 0,Peca (Rampa Sul) 0,Peca Lava 0,Peca (Rampa Sul) 0,Peca Lava 0],[Peca Lava 0,Peca (Curva Oeste) 1,Peca Recta 1,Peca (Curva Sul) 1,Peca Lava 0],[Peca Lava 0,Peca Lava 0,Peca Lava 0,Peca Lava 0,Peca Lava 0]]

m3 :: Mapa

m3 é um Mapa definido por: Mapa ((4,1),Este) [[Peca Lava 0,Peca Lava 0,Peca Lava 0,Peca Lava 0,Peca Lava 0,Peca Lava 0,Peca Lava 0],[Peca Lava 0,Peca (Curva Norte) 1,Peca (Rampa Oeste) 0,Peca Recta 0,Peca Recta 0,Peca (Curva Este) 0,Peca Lava 0],[Peca Lava 0,Peca Recta 1,Peca Lava 0,Peca Lava 0,Peca Lava 0,Peca (Rampa Sul) 0,Peca Lava 0],[Peca Lava 0,Peca (Rampa Sul) 1,Peca Lava 0,Peca (Curva Norte) 2,Peca (Rampa Oeste) 1,Peca (Curva Sul) 1,Peca Lava 0],[Peca Lava 0,Peca Recta 2,Peca Lava 0,Peca Recta 2,Peca Lava 0,Peca Lava 0,Peca Lava 0],[Peca Lava 0,Peca (Curva Oeste) 2,Peca Recta 2,Peca (Curva Sul) 2,Peca Lava 0,Peca Lava 0,Peca Lava 0],[Peca Lava 0,Peca Lava 0,Peca Lava 0,Peca Lava 0,Peca Lava 0,Peca Lava 0,Peca Lava 0]]

m4 :: Mapa

m4 é um Mapa definido por: Mapa ((2,1),Este) [[Peca Lava 0,Peca Recta 0,Peca Lava 0,Peca Lava 0,Peca Lava 0],[Peca Lava 0,Peca (Curva Norte) 0,Peca Recta 0,Peca (Curva Este) 0,Peca Lava 0],[Peca Lava 0,Peca (Rampa Sul) 0,Peca Lava 0,Peca (Rampa Sul) 0,Peca Lava 0],[Peca Lava 0,Peca (Curva Oeste) 1,Peca Recta 1,Peca (Curva Sul) 1,Peca Lava 0],[Peca Lava 0,Peca Lava 0,Peca Lava 0,Peca Lava 0]]

tab :: Mapa -> Tabuleiro

A função tab devolve o Tabuleiro de um Mapa.

Exemplo:

>>> tab m
[[Peca Lava 0,Peca Lava 0,Peca Lava 0,Peca Lava 0,Peca Lava 0],[Peca Lava 0,Peca (Curva Norte) 0,Peca Recta 0,Peca (Curva Este) 0,Peca Lava 0],[Peca Lava 0,Peca (Rampa Sul) 0,Peca Lava 0,Peca (Rampa Sul) 0,Peca Lava 0],[Peca Lava 0,Peca (Curva Oeste) 1,Peca Recta 1,Peca (Curva Sul) 1,Peca Lava 0],[Peca Lava 0,Peca Lava 0,Peca Lava 0,Peca Lava 0,Peca Lava 0]]

trajeto :: Mapa -> [(Peca, Posicao)] -> [(Peca, Posicao)]

A função trajeto devolve uma lista de pares de (Peca, Posicao),que contem a primeira passagem em cada peça do trajeto, recebendo um Mapa e uma lista de pares de (Peca, Posicao), que começa vazia.

Exemplo:

>>> trajeto m []
[(Peca Recta 0,(2,1)),(Peca (Curva Este) 0,(3,1)),(Peca (Rampa Sul) 0,(3,2)),(Peca (Curva Sul) 1,(3,3)),(Peca Recta 1,(2,3)),(Peca (Curva Oeste) 1,(1,3)),(Peca (Rampa Sul) 0,(1,2)),(Peca (Curva Norte) 0,(1,1))]

direction :: Mapa -> [(Peca, Posicao, Orientacao)] -> [(Peca, Posicao, Orientacao)]

A função direction devolve uma lista de triplos de (Peca, Posicao, Orientacao),que contem a primeira passagem em cada peça do trajeto, recebendo um Mapa e uma lista de pares de (Peca, Posicao, Orientacao), que começa vazia.

Exemplo:

>>> direction m []
[(Peca Recta 0,(2,1),Este),(Peca (Curva Este) 0,(3,1),Este),(Peca (Rampa Sul) 0,(3,2),Sul),(Peca (Curva Sul) 1,(3,3),Sul),(Peca Recta 1,(2,3),Oeste),(Peca (Curva Oeste) 1,(1,3),Oeste),(Peca (Rampa Sul) 0,(1,2),Norte),(Peca (Curva Norte) 0,(1,1),Norte)]

same_ori :: [(Peca, Posicao, Orientacao)] -> Bool

A função same_ori devolve um Bool, que depende do facto de a primeira peça do percurso e a última terem a mesma orientação.

Baseia-se no facto de a última peça ir ter à primeira, ou seja, completar o percurso.

Exemplo:

>>> same_ori (direction m [])
True

avancar2 :: (Peca, Posicao, Orientacao) -> Posicao

A função avancar2 devolve uma Posicao que é obtida através do avanço da posição que recebe, consoante orientação e tipo de Peça.

Pode ter uma peça do tipo Lava, que depois seria usada na função hasnt_lava.

Exemplo:

>>> avancar2 (Peca Recta 0, (2,1), Este)
(3,1)

position :: (Peca, Posicao, Orientacao) -> Posicao

A função position recebe um triplo de (Peca, Posicao, Orientacao) e devolve a sua Posicao.

Exemplo:

>>> position (Peca Recta 0, (2,1), Este)
(2,1)

ori :: (Peca, Posicao, Orientacao) -> Orientacao

A função ori recebe um triplo de (Peca, Posicao, Orientacao) e devolve a sua Orientacao.

Exemplo:

>>> position (Peca Recta 0, (2,1), Este)
Este

posi :: Tabuleiro -> Posicao -> Orientacao -> Peca

A função posi recebe um Tabuleiro, uma Posicao e uma Orientacao e devolve a Peca nessa posicao do Tabuleiro.

Exemplo:

>>> posi (tab m) (2,1) Este
Peca Recta 0

posi2 :: Tabuleiro -> Posicao -> Peca

A função posi2 recebe um Tabuleiro e uma Posicao e devolve a Peca nessa posicao do Tabuleiro.

Exemplo:

>>> posi2 (tab m) (3,1)
Peca (Curva Este) 0

find' :: [Peca] -> Int -> Int -> Peca

A função find' recebe uma lista de Peca, ou seja, uma linha, e dois Int, sendo o primeiro a coluna que queremos e o segundo um contador.

Devolve a Peca existente na coluna pretendida.

Exemplo:

>>> find' (head(tail (tab m))) 2 0
Peca Recta 0

tipo :: Peca -> Tipo

A função tipo recebe uma Peca e devolve o seu Tipo.

Exemplo:

>>> tipo (Peca Recta 0)
Recta

valida_alturas :: [(Peca, Posicao, Orientacao)] -> Int -> Int -> Bool

A função valida_alturas recebe uma lista de triplos de (Peca, Posicao, Orientacao) e dois Int, o primeiro relativo à altura inicial e o segundo serve para ir atualizando a altura, devolvendo um Bool consoante a veracidade das alturas.

Exemplo:

>>> valida_alturas (direction m []) 0 0
True

soPecas :: [(Peca, Posicao)] -> [Peca]

A função so_pecas recebe uma lista de pares de (Peca, Posicao) e devolve uma lista com, apenas, as Peca.

Exemplo:

>>> soPecas (trajeto m [])
[Peca Recta 0,Peca (Curva Este) 0,Peca (Rampa Sul) 0,Peca (Curva Sul) 1,Peca Recta 1,Peca (Curva Oeste) 1,Peca (Rampa Sul) 0,Peca (Curva Norte) 0]

hasnt_lava :: [(Peca, Posicao)] -> Bool

A função hasnt_lava recebe uma lista de pares de (Peca, Posicao) e devolve um Bool, consoante a existência de peças do tipo Lava na lista.

Exemplo:

>>> hasnt_lava (trajeto m [])
True

alt_matrix :: Tabuleiro -> Bool

A função alt_matrix recebe um Tabuleiro e verifica se este é só constituído por peças de Lava, com altura 0, devolvendo um Bool.

Exemplo:

>>> alt_matrix (replicate 5 (replicate 5 (Peca Lava 0)))
True

alt_line :: [Peca] -> Bool

A função alt_line recebe uma lista de Peca e verifica se este é só constituído por peças de Lava, com altura 0, devolvendo um Bool.

Exemplo:

>>> alt_line (replicate 5 (Peca Lava 0))
True

put_lava :: [Peca] -> Int -> Posicao -> [Peca]

A função put_lava recebe uma lista de Peca, um Int (contador) e uma Posicao e coloca na posicao pretendida uma peca de Lava, com altura 0.

Exemplo:

>>> put_lava (replicate 5 (Peca Recta 0)) 0 (2,0)
[Peca Recta 0,Peca Recta 0,Peca Lava 0,Peca Recta 0,Peca Recta 0]

puts :: Tabuleiro -> [Posicao] -> Tabuleiro

A função puts recebe um Tabuleiro e uma lista de Posicao e coloca nas posicões pretendida uma peca de Lava, com altura 0.

Exemplo:

>>> puts (tab m) (just_posi (trajeto m []))
[[Peca Lava 0,Peca Lava 0,Peca Lava 0,Peca Lava 0,Peca Lava 0],[Peca Lava 0,Peca Lava 0,Peca Lava 0,Peca Lava 0,Peca Lava 0],[Peca Lava 0,Peca Lava 0,Peca Lava 0,Peca Lava 0,Peca Lava 0],[Peca Lava 0,Peca Lava 0,Peca Lava 0,Peca Lava 0,Peca Lava 0],[Peca Lava 0,Peca Lava 0,Peca Lava 0,Peca Lava 0,Peca Lava 0]]

aux :: Tabuleiro -> Int -> Posicao -> Tabuleiro

A função aux recebe um Tabuleiro, um Int, que serve como contador da linha, uma Posicao e descobre em qual linha do Tabuleiro vai chamar a função put_lava.

Exemplo:

>>> aux (tab m) 0 (2,1)
[[Peca Lava 0,Peca Lava 0,Peca Lava 0,Peca Lava 0,Peca Lava 0],[Peca Lava 0,Peca (Curva Norte) 0,Peca Lava 0,Peca (Curva Este) 0,Peca Lava 0],[Peca Lava 0,Peca (Rampa Sul) 0,Peca Lava 0,Peca (Rampa Sul) 0,Peca Lava 0],[Peca Lava 0,Peca (Curva Oeste) 1,Peca Recta 1,Peca (Curva Sul) 1,Peca Lava 0],[Peca Lava 0,Peca Lava 0,Peca Lava 0,Peca Lava 0,Peca Lava 0]]

just_posi :: [(Peca, Posicao)] -> [Posicao]

A função just_posi recebe uma lista de (Peca, Posicao) e devolve uma lista com apenas as posições.

Exemplo:

>>> just_posi (trajeto m [])
[(2,1),(3,1),(3,2),(3,3),(2,3),(1,3),(1,2),(1,1)]

new_pos :: Mapa -> Mapa

A função new_pos recebe um Mapa e retorna outro Mapa, depois de avançar uma posição consoante a posição inicial, peça e orientação.

Exemplo:

>>> new_pos m
Mapa ((3,1),Este) [[Peca Lava 0,Peca Lava 0,Peca Lava 0,Peca Lava 0,Peca Lava 0],[Peca Lava 0,Peca (Curva Norte) 0,Peca Recta 0,Peca (Curva Este) 0,Peca Lava 0],[Peca Lava 0,Peca (Rampa Sul) 0,Peca Lava 0,Peca (Rampa Sul) 0,Peca Lava 0],[Peca Lava 0,Peca (Curva Oeste) 1,Peca Recta 1,Peca (Curva Sul) 1,Peca Lava 0],[Peca Lava 0,Peca Lava 0,Peca Lava 0,Peca Lava 0,Peca Lava 0]]

avancar :: Posicao -> Orientacao -> Posicao

A função avancar é utilizada na função aux e, recebendo uma Posicao e uma Orientacao, devolve a posicao para onde o carro se desloca em seguida.

>>> avancar (2,1) Este
(3,1)

muda_esq :: Orientacao -> Orientacao

A função muda_esq é utilizada na função aux e, recebendo uma Orientacao, devolve a nova orientação após uma curva à esquerda.

>>> muda_esq Norte
Oeste

muda_dir :: Orientacao -> Orientacao

A função muda_esq é utilizada na função aux e, recebendo uma Orientacao, devolve a nova orientação após uma curva à direita.

>>> muda_dir Norte
Este

valida_lava :: Mapa -> Bool

A função valida_lava verifica se a moldura de um Mapa é constituída por peças de Lava com altura 0.

>>> valida_lava m
True

aux2 :: [[Peca]] -> Bool

A função aux2 verifica se o primeiro e o último elemento de cada linha da matriz é do tipo Lava, exceto da primeira e última linha.

>>> aux2 (tab m)
True

valida_retangulo :: Mapa -> Bool

A função valida_retangulo verifica se todas as linhas do Mapa têm o mesmo número de colunas.

>>> valida_retangulo m
True

valida :: Mapa -> Bool

A função valida verifica se um mapa é válido, usando todas as funções de validção presentes neste documento.

>>> valida m2
False

>>> valida m
True

alt :: [(Peca, Posicao, Orientacao)] -> Int

A função alt devolve a altura da peca na cabeça da lista de (Peca, Posicao, Orientacao).

>>> alt (direction  m [])
0

first :: (Peca, Posicao, Orientacao) -> Peca

A função first devolve a Peca de um triplo (Peca, Posicao, Orientacao).

>>> first (Peca (Rampa Sul) 0,(1,2),Norte)
Peca (Rampa Sul) 0

aux4 :: Mapa -> Bool

A função aux4 valida as alturas do percurso de um Mapa.

Tendo em consideração o caso especil em que começa numa descida.

>>> aux4 m
True