Tarefa3_2017li1g86

Description

Módulo que implementa as funções da Tarefa 3 de Laboratórios de Informática 1.

Documentation

testesT3 :: [(Tabuleiro, Tempo, Carro)]

testesT3 é um lista de triplos (Tabuleiro, Tempo, Carro) definido por: [(t, 2, carro), (t, 1, carro2), (t, 0.7, carro2), (t, 0.7, carro3), (t, 0.9, carro3), (t, 0.5, carro), (t, 0.6, carro)]

t :: Tabuleiro

t é um Tabuleiro definido por: [[Peca Lava 0,Peca Lava 0,Peca Lava 0,Peca Lava 0,Peca Lava 0],[Peca Lava 0,Peca (Curva Norte) 0,Peca Recta 0,Peca (Curva Este) 0,Peca Lava 0],[Peca Lava 0,Peca (Rampa Sul) 0,Peca Lava 0,Peca (Rampa Sul) 0,Peca Lava 0],[Peca Lava 0,Peca (Curva Oeste) 1,Peca Recta 1,Peca (Curva Sul) 1,Peca Lava 0],[Peca Lava 0,Peca Lava 0,Peca Lava 0,Peca Lava 0,Peca Lava 0]]

carro_3 :: Carro

carro_3 é um Carro definido por: Carro {posicao = (2.5,1.5),direcao = 45,velocidade = (0,-1)}

carro_3_2 :: Carro

carro_3_2 é um Carro definido por: Carro {posicao = (2.5,1.5),direcao = 0,velocidade = (1,0)}

carro_3_3 :: Carro

carro_3_3 é um Carro definido por: Carro {posicao = (2.4,1.4),direcao = 0,velocidade = (1,0)}

carro_3_4 :: Carro

carro_3_4 é um Carro definido por: Carro {posicao = (3.4,1.7),direcao = 0,velocidade = (1,0)}

tipo :: Peca -> Tipo

A função tipo recebe uma Peca e devolve o seu Tipo.

Exemplo:

>>> tipo (Peca Lava 0)
Lava

altura :: Peca -> Int

A função altura recebe uma Peca e devolve a sua altura.

Exemplo:

>>> tipo (Peca Lava 0)
0

nextPosition :: Carro -> Tempo -> Ponto

A função nextPosition recebe um Carro e um Tempo, devolvendo a sua nova Posicao.

Exemplo:

>>> nextPosition (Carro {posicao = (1.5, 1.5), direcao = 0, velocidade = (1,0)}) 1
(2.5,1.5)

realCoords :: Ponto -> Posicao

A função realCoords recebe um Ponto e devolve a sua Posicao, ou seja, um par de inteiros.

Exemplo:

>>> realCoords (2.5,3.3)
(2,3)

posi :: Tabuleiro -> Ponto -> Peca

A função posi recebe um Tabuleiro e um Ponto, devolvendo a Peca na posição pretendida.

Exemplo:

>>> posi t2 (2.3,3.3)
Peca Recta 1

find' :: [Peca] -> Int -> Int -> Peca

A função find' recebe uma lista de Peca, ou seja, uma linha, e dois Int, sendo o primeiro a coluna que queremos e o segundo um contador.

Devolve a Peca existente na coluna pretendida.

Exemplo:

>>> find' (head(tail (tab m))) 2 0
Peca Recta 0

curva :: Peca -> Carro -> Maybe Carro

A função curva recebe uma Peca e um Carro, devolvendo um possivel novo Carro, consoante a validade do mesmo dentro de uma curva.

Exemplo:

>>> curva (Peca (Curva Este) 0) carro4
Nothing

movimenta :: Tabuleiro -> Tempo -> Carro -> Maybe Carro

A função movimenta recebe um Tabuleiro, um Tempo e um Carro, devolvendo o possível novo carro apos o movimento.

Exemplo:

>>> movimenta t2 1 carro2
Just (Carro {posicao = (3.5,1.5), direcao = 0.0, velocidade = (1.0,0.0)})

subida :: Tempo -> Peca -> Peca -> Carro -> Maybe Carro

A função subida compara as alturas quando perante situações que envolvam rampas.