Tarefa4_2017li1g86

Description

Módulo para a realização da Tarefa 4 de LI1 em 2017/18.

Documentation

testesT4 :: [(Tempo, Jogo, Acao)]

O testes a serem considerados pelo sistema de feedback para a função atualiza.

ver_propriedades_katrito :: Propriedades -> Double

A função ver_propriedades_katrito descobre a Propriedade k_atrito.

forca_atrito :: Double -> Carro -> Double -> (Double, Double)

A função forca_atrito atualiza o vetor da velocidade considerando o k_atrito, com sentido oposto ao da velocidade.

Exemplo:

>>> forca_atrito 2 c 0.3
(-1.494,-0.0)

angtorad :: Double -> Double

A função angtorad transforma um ângulo em graus, em radianos.

Exemplo:

>>> angtorad 90
1.5707963267948966

normaliza :: Double -> Double

A função normaliza mantém o ângulo entre 0º e 360º.

Exemplo:

>>> normaliza 370
10.0

ver_propriedades_kacel :: Propriedades -> Double

A função ver_propriedades_kacel descobre a Propriedade k_acel.

forca_aceleracao :: Double -> Double -> Double -> (Double, Double)

A função forca_aceleracao atualiza o vetor da velocidade considerando que o carro está a acelerar, com sentido e direção do carro.

Exemplo:

>>> forca_aceleracao 2 45  0.3
(0.4242640687119285,0.42426406871192845)

forca_travar :: Double -> Double -> Double -> (Double, Double)

A função forca_travar atualiza o vetor da velocidade considerando que o carro está a travar, com a mesma direção e sentido oposto ao do carro.

Exemplo:

>>> forca_travar 2 45  0.3
(-0.4242640687119285,-0.42426406871192845)

forca_ace_trav :: Acao -> Carro -> Double -> Double -> (Double, Double)

A função forca_ace_trav vê se o carro está a acelarar ou a travar. Se estiver a fazer apenas um destes, calcula o vetor da força. Caso contrário, a força não é considerada.

Exemplo:

>>> forca_ace_trav acao4 c 1 0.3
(0.22981333293569342,0.19283628290596178)

ver_propriedades_kpeso :: Propriedades -> Double

A função ver_propriedades_kpeso descobre a Propriedade k_peso.

forca_gravidade :: Mapa -> Carro -> Double -> (Double, Double)

A função forca_gravidade vê se o carro está numa rampa, descobrindo de que forma esta afeta o vetor velocidade do Carro, assumindo que as rampas têm sempre inclinação 45º.

diferenca_angulos :: Double -> Double -> Double

A função diferenca_angulos verifica a diferença entre o ângulo da direção e o ângulo da velocidade.

Exemplo:

>>> diferenca_angulos 47 45
-2.0

angulo_total :: Double -> Double -> Double

A função angulo_total verifica de que forma k_pneus vai afetar o Carro.

angulo_velocidade :: Carro -> Double

A função angulo_velocidade calcula o ângulo que o vetor da velocidade faz.

Exemplo:

>>> angulo_velocidade c
0.0

forca_pneus :: Double -> Double -> Double -> Double -> (Double, Double)

A função forca_pneus calcula as coordenadas cartesianas da força pneus.

ver_propriedades_kpneus :: Propriedades -> Double

A função ver_propriedades_kpneus descobre a Propriedade k_pneus.

tipo_4 :: Peca -> Tipo

A função tipo_4 recebe uma Peca e devolve o seu Tipo.

Exemplo:

>>> tipo_4 (Peca Lava 0)
Lava

soma_forcas :: (Double, Double) -> (Double, Double) -> (Double, Double) -> (Double, Double) -> (Double, Double)

A função soma_forcas soma a força do atrito, da gravidade, dos pneus e da aceleração.

angulo :: Tempo -> Double -> Double -> Acao -> Double

A função angulo determina o novo ângulo do carro, consoante a rotação implicada pelo Tempo da Acao.

ver_angulo :: Carro -> Double

A função ver_angulo determina o ângulo do carro, a partir deste.

Exemplo:

>>> ver_angulo c
40.0

ver_propriedades_kroda :: Propriedades -> Double

A função ver_propriedades_kroda descobre a Propriedade k_roda.

ver2 :: Carro -> [Posicao] -> [Posicao]

A função ver2 atualiza o Histórico do Carro em questão.

Exemplo:

>>> ver2 c []
[(1,1)]

change :: [Tempo] -> Int -> Tempo -> [Tempo]

A função change altera um elemento na posição x da lista que recebe.

Exemplo:

>>> change [1,2,3,4] 1 7
[1.0,7.0,3.0,4.0]

change2 :: [Carro] -> Int -> Carro -> [Carro]

A função change2 altera um elemento na posição x da lista que recebe.

change3 :: [[Posicao]] -> Int -> [Posicao] -> [[Posicao]]

A função change3 altera um elemento na posição x da lista que recebe.

ver_propriedades_knitro :: Propriedades -> Double

A função ver_propriedades_knitro descobre a Propriedade k_nitro.

atualiza_velocidade_nitro :: Carro -> Double -> Tempo -> Carro

A função atualiza_velocidade_nitro atualiza a velocidade do Carro quando este tem um Nitro ativo sobre si.

atualiza_tempo_nitro :: Jogo -> Int -> Tempo -> Jogo

A função atualiza_tempo_nitro atualiza o tempo de Nitro restante de um Carro quando este tem um Nitro ativo sobre si.

nitroEffect :: Jogo -> Maybe Int -> Double -> Tempo -> Jogo

A função nitroEffect atualiza o Jogo com as alterações devidas ao factor Nitro.

fromJust' :: Maybe Int -> Int

A função fromJust' recebe um Maybe Int e devolve um Int, quando o input é diferente de Nothing__.

ver_nitro :: Jogo -> Acao -> Tempo -> Jogo

A função ver_nitro atualiza o Jogo consoante a existência de Nitro ativo ou não.

vector :: Carro -> (Double, Double) -> Carro

A função vector troca o vector velocidade de um Carro pelo '(Double,Double)' que recebe.

Exemplo:

>>> vector c (1,1)
Carro {posicao = (1.5,1.5), direcao = 40.0, velocidade = (1.0,1.0)}

carro_atualizado_angulo_somaforcas :: Tempo -> Carro -> (Double, Double) -> Propriedades -> Acao -> Carro

A função carro_atualizado_angulo_somaforcas junta a atualização do ângulo e das forças.

m :: Mapa

m é um Mapa definido por: Mapa ((2,1),Este) [[Peca Lava 0,Peca Lava 0,Peca Lava 0,Peca Lava 0,Peca Lava 0],[Peca Lava 0,Peca (Curva Norte) (-1),Peca Recta (-1),Peca (Curva Este) (-1),Peca Lava 0],[Peca Lava 0,Peca Recta (-1),Peca Lava 0,Peca Recta (-1),Peca Lava 0],[Peca Lava 0,Peca (Curva Oeste) (-1),Peca Recta (-1),Peca (Curva Sul) (-1),Peca Lava 0],[Peca Lava 0,Peca Lava 0,Peca Lava 0,Peca Lava 0,Peca Lava 0]]

c :: Carro

c é um Carro definido por: Carro {posicao = (1.5,1.5),direcao = 40.0,velocidade = (2.49,0.0)}

p :: Propriedades

p são Propriedades definido por: Propriedades {k_atrito = 2.0,k_pneus = 3.0,k_acel = 4.0,k_peso = 2.0,k_nitro = 15.0,k_roda = 180.0}

game1 :: Jogo

game1 é um Jogo definido por: Jogo{mapa = Mapa ((2,1),Este) [[Peca Lava 0,Peca Lava 0,Peca Lava 0,Peca Lava 0,Peca Lava 0],[Peca Lava 0,Peca (Curva Norte) (-1),Peca Recta (-1),Peca (Curva Este) (-1),Peca Lava 0],[Peca Lava 0,Peca Recta (-1),Peca Lava 0,Peca Recta (-1),Peca Lava 0],[Peca Lava 0,Peca (Curva Oeste) (-1),Peca Recta (-1),Peca (Curva Sul) (-1),Peca Lava 0],[Peca Lava 0,Peca Lava 0,Peca Lava 0,Peca Lava 0,Peca Lava 0]] ,pista = Propriedades {k_atrito = 2.0,k_pneus = 3.0,k_acel = 4.0,k_peso = 2.0,k_nitro = 15.0,k_roda = 180.0} ,carros = [Carro {posicao = (1.5,1.5),direcao = 40.0,velocidade = (2.49,0.0)} ,Carro {posicao = (2.5,1.5),direcao = 0.0,velocidade = (52.6,0.0)}], nitros = [1.0,2.0],historico = [[],[(2,2)]]}

game2 :: Jogo

game2 é um Jogo definido por: Jogo{mapa = Mapa ((2,1),Este) [[Peca Lava 0,Peca Lava 0,Peca Lava 0,Peca Lava 0,Peca Lava 0],[Peca Lava 0,Peca (Curva Norte) (-1),Peca Recta (-1),Peca (Curva Este) (-1),Peca Lava 0],[Peca Lava 0,Peca Recta (-1),Peca Lava 0,Peca Recta (-1),Peca Lava 0],[Peca Lava 0,Peca (Curva Oeste) (-1),Peca Recta (-1),Peca (Curva Sul) (-1),Peca Lava 0],[Peca Lava 0,Peca Lava 0,Peca Lava 0,Peca Lava 0,Peca Lava 0]] ,pista = Propriedades {k_atrito = 2.0,k_pneus = 3.0,k_acel = 4.0,k_peso = 2.0,k_nitro = 15.0,k_roda = 180.0} ,carros = [Carro {posicao = (1.5,1.5),direcao = 40.0,velocidade = (2.5,0.0)} ,Carro {posicao = (2.5,1.5),direcao = 0.0,velocidade = (1.0,0.0)}], nitros = [1.0,1.0],historico = [[],[(2,2)]]}

game3 :: Jogo

game3 é um Jogo definido por: Jogo{mapa = Mapa ((2,1),Este) [[Peca Lava 0,Peca Lava 0,Peca Lava 0,Peca Lava 0,Peca Lava 0],[Peca Lava 0,Peca (Curva Norte) (-1),Peca Recta (-1),Peca (Curva Este) (-1),Peca Lava 0],[Peca Lava 0,Peca Recta (-1),Peca Lava 0,Peca Recta (-1),Peca Lava 0],[Peca Lava 0,Peca (Curva Oeste) (-1),Peca Recta (-1),Peca (Curva Sul) (-1),Peca Lava 0],[Peca Lava 0,Peca Lava 0,Peca Lava 0,Peca Lava 0,Peca Lava 0]] ,pista = Propriedades {k_atrito = 2.0,k_pneus = 3.0,k_acel = 4.0,k_peso = 2.0,k_nitro = 15.0,k_roda = 180.0} ,carros = [Carro {posicao = (1.5,1.5),direcao = 40.0,velocidade = (2.5,0.0)}], nitros = [0.0],historico = [[]]}

game4 :: Jogo

game4 é um Jogo definido por: Jogo{mapa = Mapa ((2,1),Este) [[Peca Lava 0,Peca Lava 0,Peca Lava 0,Peca Lava 0,Peca Lava 0],[Peca Lava 0,Peca (Curva Norte) (-1),Peca Recta (-1),Peca (Curva Este) (-1),Peca Lava 0],[Peca Lava 0,Peca Recta (-1),Peca Lava 0,Peca Recta (-1),Peca Lava 0],[Peca Lava 0,Peca (Curva Oeste) (-1),Peca Recta (-1),Peca (Curva Sul) (-1),Peca Lava 0],[Peca Lava 0,Peca Lava 0,Peca Lava 0,Peca Lava 0,Peca Lava 0]] ,pista = Propriedades {k_atrito = 2.0,k_pneus = 3.0,k_acel = 4.0,k_peso = 2.0,k_nitro = 15.0,k_roda = 180.0} ,carros = [Carro {posicao = (1.5,1.5),direcao = 0.0,velocidade = (1,0.0)},Carro {posicao = (1.5,1.5),direcao = 0.0,velocidade = (2,0.0)}], nitros = [0.0,2.0],historico = [[],[(1,1)]]}

game5 :: Jogo

game5 é um Jogo definido por: Jogo{mapa = Mapa ((2,1),Este) [[Peca Lava 0,Peca Lava 0,Peca Lava 0,Peca Lava 0,Peca Lava 0],[Peca Lava 0,Peca (Curva Norte) (-1),Peca Recta (-1),Peca (Curva Este) (-1),Peca Lava 0],[Peca Lava 0,Peca Recta (-1),Peca Lava 0,Peca Recta (-1),Peca Lava 0],[Peca Lava 0,Peca (Curva Oeste) (-1),Peca Recta (-1),Peca (Curva Sul) (-1),Peca Lava 0],[Peca Lava 0,Peca Lava 0,Peca Lava 0,Peca Lava 0,Peca Lava 0]] ,pista = Propriedades {k_atrito = 2.0,k_pneus = 3.0,k_acel = 4.0,k_peso = 2.0,k_nitro = 15.0,k_roda = 180.0} ,carros = [Carro {posicao = (1.5,1.5),direcao = 10.0,velocidade = (2.5,0.0)}], nitros = [0.0],historico = [[]]}

hist :: [[Posicao]]

hist é uma '[[Posicao]]' definida por: [[],[(2,2)]].

cars :: [Carro]

cars é uma '[Carro]' definida por: [Carro {posicao = (1.5,1.5),direcao = 40.0,velocidade = (2.5,0.0)},Carro {posicao = (2.5,1.5),direcao = 0.0,velocidade = (1.0,0.0)}].

velocity :: Carro -> (Double, Double)

A função velocity devolve a velocidade de um Carro, através do mesmo.

Exemplo:

>>> velocity c
(2.49,0.0)

acao1 :: Acao

acao1 é uma Acao definido por: Acao {acelerar = False, travar = True, esquerda= False, direita= True, nitro = Nothing}

acao2 :: Acao

acao2 é uma Acao definido por: Acao {acelerar = False, travar = False, esquerda= False, direita= False, nitro = Nothing}

acao3 :: Acao

acao3 é uma Acao definido por: Acao {acelerar = False, travar = False, esquerda= False, direita= False, nitro = Just 1}

acao4 :: Acao

acao4 é uma Acao definido por: Acao {acelerar = True, travar = False, esquerda= False, direita= False, nitro = Nothing}

timer :: (Double, Double) -> Tempo -> (Double, Double)

A função timer multiplica o vetor do input pelo Tempo.

Exemplo:

>>> timer (1,1) 0.3
(0.3,0.3)

atualiza :: Tempo -> Jogo -> Int -> Acao -> Jogo

Função usada para atualizar o estado do jogo dadas as ações de um jogador num determinado período de tempo.