Tarefa6_2017li1g86

Description

Módulo para a realização da Tarefa 6 de LI1 em 2017/18.

Documentation

fromJust' :: Maybe Carro -> Carro

A função fromJust' recebe um Maybe Carro e devolve um Carro, quando o input é diferente de Nothing.

car :: Carro

car é um Carro definido por: Carro {posicao = (1,1), direcao = 0.0, velocidade = (1.0,0.0)}

bot

Arguments

:: Tempo
tempo decorrido desde a última decisão

-> Jogo
estado atual do jogo

-> Int
identificador do jogador dentro do estado

-> Acao
a decisão tomada pelo bot

Função usada para simular um bot no jogo Micro Machines. Em cada instante, dado o tempo decorrido, o estado do jogo e o identificador do jogador, toma uma ação.

see :: Carro -> Tabuleiro -> Peca

A função see recebe um Carro e um Tabuleiro, devolvendo a Peca onde o primeiro se encontra.

Exemplo:

>>> see car t
Peca (Curva Norte) 0

posi3 :: Tabuleiro -> Ponto -> Peca

A função posi3 recebe um Tabuleiro e um Ponto, devolvendo a Peca na posição pretendida.

Exemplo:

>>> posi3 t (2.3,3.3)
Peca Recta 1

find2 :: [Peca] -> Int -> Int -> Peca

A função find2 recebe uma lista de Peca, ou seja, uma linha, e dois Int, sendo o primeiro a coluna que queremos e o segundo um contador.

Devolve a Peca existente na coluna pretendida.

Exemplo:

>>> find' (head(tail (tab m))) 2 0
Peca Recta 0

chooseAction2 :: Carro -> Tempo -> Mapa -> Acao

A função chooseAction2 recebe um Carro, um Tempo e um Mapa, devolvendo a Acao que o carro deve tomar, consoante a peça onde se encontra.

Exemplo:

>>> chooseAction2 car 0.1 (Mapa ((2,1),Este) t)
Acao {acelerar = True, travar = False, esquerda = False, direita = True, nitro = Nothing}

tira :: Carro -> [(Peca, Posicao, Orientacao)] -> Orientacao

A função tira recebe um Carro e uma lista de triplos '(Peca,Posicao,Orientacao)', devolvendo a Orientação da peça onde o carro se encontra.

pick2 :: [Carro] -> Int -> Carro

A função pick2 recebe uma lista de Carro e um Int, devolvendo o Carro na posição correspondente.

Exemplo:

>>> pick2 [car,carro_3] 1
Carro {posicao = (1.0,1.0), direcao = 0.0, velocidade = (1.0,0.0)}

find_actual :: Tempo -> Carro -> Carro

A função find_actual recebe um Tempo e um Carro, devolvendo o Carro na sua possível posição atualizada.

Exemplo:

>>> find_actual 0.1 car
Carro {posicao = (1.1,1.0), direcao = 0.0, velocidade = (1.0,0.0)}

tipo2 :: Peca -> Tipo

A função tipo2 recebe uma Peca e devolve o seu Tipo.

Exemplo:

>>> tipo2 (Peca Recta 1)
Recta